requirements.txt
telethon==1.28.5
sqlalchemy==2.0.23
alembic==1.12.1
pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
asyncio==3.4.3
